requirements.txt
# requirements.txt
PyQt5
requests
python-dotenv
pandas
openpyxl